Update onnxUI.py
I tried to format the history.txt a bit better. It was really hard to select the positive and negative prompts to copy-paste so I put them in separate lines. Also "sched_name" isn't returning the right scheduler. I tried to use "scheduler" but it just prints the whole thing.