add limit to img2img adjustment if using dpms
if using a very low denoise the step count can get adjusted to way over 1000 which causes the error "ValueError: only one element tensors can be converted to Python scalars" if using dpms. This commit will cap it to 1000 only when using dpms.

There is likely a way to fix this to allow for >1000, but I'm not sure how yet.