changed default temp folder used by ui.
since gradio uses the lib tempfile to store generated files to be displayed by ui, the files are saved on OS "temp" folder.
it is not a problem in linux as how the temp folder works, but on windows it stays until user cleans it. so this modification changes the default temp folder to a more approriate, then when stopping the ui it tries to clean the folder.